soc: alpine: add udma helpers
One or more DMA controllers (called Universal DMA) are exposed as PCI
endpoints for some units connected to the I/O fabric. The DMA controller
registers are mapped in BARs.

These helpers help drivers supporting units connected to the Alpine I/O
fabric, to configure and use the exposed UDMA.

The Alpine UDMA is a full-duplex DMA consisting of a Tx (memory to
stream -m2s-) and an Rx (stream to memory -s2m-) DMA engines. Multiple
queues are available per DMA, with each a description and a completion
ring.

Signed-off-by: Antoine Tenart <antoine.tenart@free-electrons.com>